Add leaflet to display address on map
